+++ src/Pages/ProductPage/ProductPage.js
@@ -1,5 +1,6 @@
 //individual page for product that allows add to cart state change functionality
 import React from 'react';
+
 function ProductPage() {
     return (
         <div>
