+++ src/index.js
@@ -4,7 +4,7 @@ import './index.css';
 import App from './App';
 import reportWebVitals from './reportWebVitals';
 import { StateProvider } from './StateProvider';
-import reducer, { initalState } from './Products/reducer';
+import reducer, { initalState } from './Pages/reducer';
 
 const root = ReactDOM.createRoot(document.getElementById('root'));
 root.render(
